Refactor and move trend functions to separate lib file
