+++ src/datasets/summarizations/utils/time-series.ts
@@ -1,6 +1,5 @@
-import { TimeSeriesPoint, NumPoint } from '../../metas/types';
+import { TimeSeriesPoint } from '../../metas/types';
 import { DAY, WEEK } from '../../../utils/timeUnits';
-import { normalizePoints } from './commons';
 
 /**
  * Group time-series points by the week number of their x-values (date).
@@ -32,92 +31,3 @@ export function timeSeriesPointToNumPoint(point: TimeSeriesPoint) {
   };
 }
 
-export interface Cone2D {
-  startAngleRad: number;
-  endAngleRad: number;
-}
-
-export interface TimeSeriesTrend {
-  idxStart: number;
-  idxEnd: number;
-  timeStart: Date;
-  timeEnd: Date;
-  pctSpan: number;
-  cone: Cone2D;
-}
-
-
-export function normalizedUniformPartiallyLinearEpsApprox(points: TimeSeriesPoint[], eps: number): TimeSeriesTrend[] {
-  const numPoints = points.map(timeSeriesPointToNumPoint);
-  const normalizedPoints = normalizePoints(numPoints, {}, { min: 0 });
-
-  if (normalizedPoints.length <= 1) {
-    return [];
-  } else {
-    const trends: TimeSeriesTrend[] = [];
-    // Modified Sklansky and Gonzalez algorithm for extracting trends
-    let i = 0;
-    let j = 1;
-    let coneij: Cone2D;
-    while (j < normalizedPoints.length) {
-      let k = j;
-      let coneik = calcCone(normalizedPoints[i], normalizedPoints[k], eps);
-      coneij = coneik;
-      do {
-        coneij = intersectCone(coneij, coneik) as Cone2D;
-        j = k;
-        k = k + 1;
-        if (k === normalizedPoints.length) {
-          break;
-        }
-        coneik = calcCone(normalizedPoints[i], normalizedPoints[k], eps);
-      } while (intersectCone(coneij, coneik) !== null);
-
-      trends.push({
-        idxStart: i,
-        idxEnd: j,
-        timeStart: points[i].x,
-        timeEnd: points[j].x,
-        pctSpan: normalizedPoints[j].x - normalizedPoints[i].x,
-        cone: coneij,
-      });
-
-      i = j;
-      j = k;
-    }
-    return trends;
-  }
-}
-
-function calcCone(p1: NumPoint, p2: NumPoint, eps: number): Cone2D {
-  const { x: x1, y: y1 } = p1;
-  const { x: x2, y: y2 } = p2;
-  const dx = x2 - x1;
-  const dy = y2 - y1;
-  if (dx * dx + dy * dy <= eps * eps) {
-    return {
-      startAngleRad: -Math.PI / 2,
-      endAngleRad: Math.PI / 2,
-    };
-  } else {
-    const a1 = Math.atan((dx * dy - eps * Math.sqrt(dx * dx + dy * dy - eps * eps)) / (dx * dx - eps * eps));
-    const a2 = Math.atan((dx * dy + eps * Math.sqrt(dx * dx + dy * dy - eps * eps)) / (dx * dx - eps * eps));
-    return {
-      startAngleRad: Math.min(a1, a2),
-      endAngleRad: Math.max(a1, a2),
-    };
-  }
-}
-
-function intersectCone(c1: Cone2D, c2: Cone2D): Cone2D | null {
-  const startAngleRad = Math.max(c1.startAngleRad, c2.startAngleRad);
-  const endAngleRad = Math.min(c1.endAngleRad, c2.endAngleRad);
-  if (startAngleRad <= endAngleRad) {
-    return {
-      startAngleRad,
-      endAngleRad,
-    };
-  } else {
-    return null;
-  }
-}
